pyproject.toml
[tool.poetry]
name = "nonebot-plugin-jx3"
version = "1.0.5"
description = "基于 NoneBot2 的剑网三全功能 QQ 群助手插件"
authors = ["water <1594567228@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_jx3"}]
exclude = ["nonebot_plugin_jx3/data/**/*"]
homepage = "https://github.com/fuyang0811/nonebot-plugin-jx3"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = ">=2.0.0"
nonebot-adapter-onebot = ">=2.0.0"
nonebot-plugin-apscheduler = ">=0.3.0"
nonebot-plugin-htmlrender = ">=0.2.0"
nonebot-plugin-localstore = ">=0.6.0"
httpx = ">=0.19.0"
aiohttp = ">=3.8.0"
jinja2 = ">=3.0.0"
pydantic = ">=2.0.0"
tinydb = ">=4.5.1"
matplotlib = ">=3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
